Create basic structure for Missions
