+++ src/App.js
@@ -1,32 +1,26 @@
 import React from 'react';
+import { Redirect, Route, Switch } from 'react-router';
 import './App.css';
-import { Switch, Route, Redirect } from 'react-router';
-import ProfileComponent from './components/Profile';
-import Rockets from './components/Rockets';
 import Header from './components/Header';
 import Missions from './components/Missions';
+import ProfileComponent from './components/Profile';
+import Rockets from './components/Rockets';
 
 const App = () => (
   <>
     <Header />
     <Switch>
-
-    <Route exact path="/">
-        <Rockets/>
+      <Route exact path="/">
+        <Rockets />
       </Route>
-
-      <Route path="/profile">
-        <ProfileComponent/>
-      </Route>
-
       <Route path="/missions">
-        <Missions/>
+        <Missions />
       </Route>
-        
+      <Route path="/profile">
+        <ProfileComponent />
+      </Route>
       <Redirect from="*" to="/" />
-
-      </ Switch>
-    
+    </Switch>
   </>
 );
