+++ src/redux/store.js
@@ -1,6 +1,13 @@
-import { createStore } from 'redux';
-import reducer from './ducks/counter';
+import { applyMiddleware, combineReducers, createStore } from 'redux';
+import logger from 'redux-logger';
+import thunk from 'redux-thunk';
+import missionsReducer from './ducks/missions';
 
-const store = createStore(reducer);
+const reducer = combineReducers({
+  missionsReducer,
+  // additional reducers could be added here
+});
+
+const store = createStore(reducer, applyMiddleware(logger, thunk));
 
 export default store;
